docs/coverage/node/html/test/wpt/xhr/send-redirect-post-upload.js.gcov.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">

<html lang="en">

<head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
  <title>LCOV - lcov.info - test/wpt/xhr/send-redirect-post-upload.js</title>
  <link rel="stylesheet" type="text/css" href="../../../gcov.css">
</head>

<body>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="title">LCOV - code coverage report</td></tr>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>

    <tr>
      <td width="100%">
        <table cellpadding=1 border=0 width="100%">
          <tr>
            <td width="10%" class="headerItem">Current view:</td>
            <td width="35%" class="headerValue"><a href="../../../index.html">top level</a> - <a href="index.html">test/wpt/xhr</a> - send-redirect-post-upload.js<span style="font-size: 80%;"> (source / <a href="send-redirect-post-upload.js.func-sort-c.html">functions</a>)</span></td>
            <td width="5%"></td>
            <td width="15%"></td>
            <td width="10%" class="headerCovTableHead">Hit</td>
            <td width="10%" class="headerCovTableHead">Total</td>
            <td width="15%" class="headerCovTableHead">Coverage</td>
          </tr>
          <tr>
            <td class="headerItem">Test:</td>
            <td class="headerValue">lcov.info</td>
            <td></td>
            <td class="headerItem">Lines:</td>
            <td class="headerCovTableEntry">285</td>
            <td class="headerCovTableEntry">287</td>
            <td class="headerCovTableEntryHi">99.3 %</td>
          </tr>
          <tr>
            <td class="headerItem">Date:</td>
            <td class="headerValue">2025-01-24 02:02:48</td>
            <td></td>
            <td class="headerItem">Functions:</td>
            <td class="headerCovTableEntry">9</td>
            <td class="headerCovTableEntry">9</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td></td>
            <td></td>
            <td></td>
            <td class="headerItem">Branches:</td>
            <td class="headerCovTableEntry">19</td>
            <td class="headerCovTableEntry">20</td>
            <td class="headerCovTableEntryHi">95.0 %</td>
          </tr>
          <tr><td><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
        </table>
      </td>
    </tr>

    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
  </table>

  <table cellpadding=0 cellspacing=0 border=0>
    <tr>
      <td><br></td>
    </tr>
    <tr>
      <td>
<pre class="sourceHeading">           Branch data     Line data    Source code</pre>
<pre class="source">
<a name="1"><span class="lineNum">       1 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : import assert from 'node:assert/strict';</span></a>
<a name="2"><span class="lineNum">       2 </span>                :<span class="lineCov">          1 : import { XMLHttpRequest } from '../../../lib/whatwg-xhr.js';</span></a>
<a name="3"><span class="lineNum">       3 </span>                :<span class="lineCov">          1 : </span></a>
<a name="4"><span class="lineNum">       4 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : export default (activeURL) =&gt; {</span></a>
<a name="5"><span class="lineNum">       5 </span>                :<span class="lineCov">          1 : </span></a>
<a name="6"><span class="lineNum">       6 </span>                :<span class="lineCov">          1 :         const stringBody = 'Test Message'.repeat(1000);</span></a>
<a name="7"><span class="lineNum">       7 </span>                :<span class="lineCov">          1 :         const blobBody = new Blob(new Array(1000).fill('Test Message'));</span></a>
<a name="8"><span class="lineNum">       8 </span>                :<span class="lineCov">          1 : </span></a>
<a name="9"><span class="lineNum">       9 </span>                :<span class="lineCov">          1 :         testRedirectPost( { name: '301',          code: 301, expectResendPost: false, body: stringBody} );</span></a>
<a name="10"><span class="lineNum">      10 </span>                :<span class="lineCov">          1 :         testRedirectPost( { name: '302',          code: 302, expectResendPost: false, body: stringBody} );</span></a>
<a name="11"><span class="lineNum">      11 </span>                :<span class="lineCov">          1 :         testRedirectPost( { name: '303',          code: 303, expectResendPost: false, body: stringBody} );</span></a>
<a name="12"><span class="lineNum">      12 </span>                :<span class="lineCov">          1 :         testRedirectPost( { name: '307 (string)', code: 307, expectResendPost: true,  body: stringBody, expectedBody: stringBody } );</span></a>
<a name="13"><span class="lineNum">      13 </span>                :<span class="lineCov">          1 :         testRedirectPost( { name: '307 (blob)',   code: 307, expectResendPost: true,  body: blobBody,   expectedBody: stringBody, expectedContentType: 'NO' } );</span></a>
<a name="14"><span class="lineNum">      14 </span>                :<span class="lineCov">          1 : </span></a>
<a name="15"><span class="lineNum">      15 </span>           [<span class="branchCov" title="Branch 0 was taken 5 times"> + </span>]:<span class="lineCov">          1 :         function testRedirectPost(params){</span></a>
<a name="16"><span class="lineNum">      16 </span>                :<span class="lineCov">          5 :                 let actual = [];</span></a>
<a name="17"><span class="lineNum">      17 </span>                :<span class="lineCov">          5 :                 // We check upload.onprogress with a boolean because it *might* fire more than once</span></a>
<a name="18"><span class="lineNum">      18 </span>                :<span class="lineCov">          5 :                 let progressFiredReadyState1 = false;</span></a>
<a name="19"><span class="lineNum">      19 </span>                :<span class="lineCov">          5 : </span></a>
<a name="20"><span class="lineNum">      20 </span>                :<span class="lineCov">          5 :                 let expectedHeaders, expectedEvents;</span></a>
<a name="21"><span class="lineNum">      21 </span>                :<span class="lineCov">          5 : </span></a>
<a name="22"><span class="lineNum">      22 </span>                :<span class="lineCov">          5 :                 // 307 redirects should resend the POST data, and events and headers will be a little different..</span></a>
<a name="23"><span class="lineNum">      23 </span>           [<span class="branchCov" title="Branch 0 was taken 2 times"> + </span>]:<span class="lineCov">          5 :                 if(params.expectResendPost){</span></a>
<a name="24"><span class="lineNum">      24 </span>                :<span class="lineCov">          2 :                         expectedHeaders = {</span></a>
<a name="25"><span class="lineNum">      25 </span>                :<span class="lineCov">          2 :                                 'X-Request-Content-Length': '12000',</span></a>
<a name="26"><span class="lineNum">      26 </span>                :<span class="lineCov">          2 :                                 'X-Request-Content-Type': 'text/plain;charset=UTF-8',</span></a>
<a name="27"><span class="lineNum">      27 </span>                :<span class="lineCov">          2 :                                 'X-Request-Method': 'POST',</span></a>
<a name="28"><span class="lineNum">      28 </span>                :<span class="lineCov">          2 :                                 'X-Request-Query': 'NO',</span></a>
<a name="29"><span class="lineNum">      29 </span>                :<span class="lineCov">          2 :                                 'Content-Length': '12000'</span></a>
<a name="30"><span class="lineNum">      30 </span>                :<span class="lineCov">          2 :                         };</span></a>
<a name="31"><span class="lineNum">      31 </span>                :<span class="lineCov">          2 : </span></a>
<a name="32"><span class="lineNum">      32 </span>                :<span class="lineCov">          2 :                         expectedEvents = [</span></a>
<a name="33"><span class="lineNum">      33 </span>                :<span class="lineCov">          2 :                                 'xhr onreadystatechange 1',</span></a>
<a name="34"><span class="lineNum">      34 </span>                :<span class="lineCov">          2 :                                 'xhr loadstart 1',</span></a>
<a name="35"><span class="lineNum">      35 </span>                :<span class="lineCov">          2 :                                 'upload loadstart 1',</span></a>
<a name="36"><span class="lineNum">      36 </span>                :<span class="lineCov">          2 :                                 'upload loadend 1',</span></a>
<a name="37"><span class="lineNum">      37 </span>                :<span class="lineCov">          2 :                                 'xhr onreadystatechange 2',</span></a>
<a name="38"><span class="lineNum">      38 </span>                :<span class="lineCov">          2 :                                 'xhr onreadystatechange 3',</span></a>
<a name="39"><span class="lineNum">      39 </span>                :<span class="lineCov">          2 :                                 'xhr onreadystatechange 4',</span></a>
<a name="40"><span class="lineNum">      40 </span>                :<span class="lineCov">          2 :                                 'xhr load 4',</span></a>
<a name="41"><span class="lineNum">      41 </span>                :<span class="lineCov">          2 :                                 'xhr loadend 4'</span></a>
<a name="42"><span class="lineNum">      42 </span>                :<span class="lineCov">          2 :                         ];</span></a>
<a name="43"><span class="lineNum">      43 </span>           [<span class="branchCov" title="Branch 0 was taken 3 times"> + </span>]:<span class="lineCov">          2 :                 }</span></a>
<a name="44"><span class="lineNum">      44 </span>                :<span class="lineCov">          3 :                 else {</span></a>
<a name="45"><span class="lineNum">      45 </span>                :<span class="lineCov">          3 :                         // setting the right expectations for POST resent as GET without request body</span></a>
<a name="46"><span class="lineNum">      46 </span>                :<span class="lineCov">          3 :                         expectedHeaders = {</span></a>
<a name="47"><span class="lineNum">      47 </span>                :<span class="lineCov">          3 :                                 'X-Request-Content-Length': 'NO',</span></a>
<a name="48"><span class="lineNum">      48 </span>                :<span class="lineCov">          3 :                                 'X-Request-Content-Type': 'NO',</span></a>
<a name="49"><span class="lineNum">      49 </span>                :<span class="lineCov">          3 :                                 'X-Request-Method': 'GET',</span></a>
<a name="50"><span class="lineNum">      50 </span>                :<span class="lineCov">          3 :                                 'X-Request-Query': 'NO'</span></a>
<a name="51"><span class="lineNum">      51 </span>                :<span class="lineCov">          3 :                         };</span></a>
<a name="52"><span class="lineNum">      52 </span>                :<span class="lineCov">          3 : </span></a>
<a name="53"><span class="lineNum">      53 </span>                :<span class="lineCov">          3 :                         expectedEvents = [</span></a>
<a name="54"><span class="lineNum">      54 </span>                :<span class="lineCov">          3 :                                 'xhr onreadystatechange 1',</span></a>
<a name="55"><span class="lineNum">      55 </span>                :<span class="lineCov">          3 :                                 'xhr loadstart 1',</span></a>
<a name="56"><span class="lineNum">      56 </span>                :<span class="lineCov">          3 :                                 'upload loadstart 1',</span></a>
<a name="57"><span class="lineNum">      57 </span>                :<span class="lineCov">          3 :                                 'upload loadend 1',</span></a>
<a name="58"><span class="lineNum">      58 </span>                :<span class="lineCov">          3 :                                 'xhr onreadystatechange 2',</span></a>
<a name="59"><span class="lineNum">      59 </span>                :<span class="lineCov">          3 :                                 // we expect no onreadystatechange readyState=3 event because there is no loading content</span></a>
<a name="60"><span class="lineNum">      60 </span>                :<span class="lineCov">          3 :                                 'xhr onreadystatechange 4',</span></a>
<a name="61"><span class="lineNum">      61 </span>                :<span class="lineCov">          3 :                                 'xhr load 4',</span></a>
<a name="62"><span class="lineNum">      62 </span>                :<span class="lineCov">          3 :                                 'xhr loadend 4'</span></a>
<a name="63"><span class="lineNum">      63 </span>                :<span class="lineCov">          3 :                         ];</span></a>
<a name="64"><span class="lineNum">      64 </span>                :<span class="lineCov">          3 :                 }</span></a>
<a name="65"><span class="lineNum">      65 </span>                :<span class="lineCov">          5 :                 // Override expectations if provided.</span></a>
<a name="66"><span class="lineNum">      66 </span>                :<span class="lineCov">          5 :                 if(params.expectedContentType)</span></a>
<a name="67"><span class="lineNum">      67 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          5 :                         expectedHeaders['X-Request-Content-Type'] = params.expectedContentType;</span></a>
<a name="68"><span class="lineNum">      68 </span>                :<span class="lineCov">          5 : </span></a>
<a name="69"><span class="lineNum">      69 </span>                :<span class="lineCov">          5 :                 let xhr = new XMLHttpRequest();</span></a>
<a name="70"><span class="lineNum">      70 </span>                :<span class="lineCov">          5 : </span></a>
<a name="71"><span class="lineNum">      71 </span>           [<span class="branchCov" title="Branch 0 was taken 5 times"> + </span>]:<span class="lineCov">          5 :                 xhr.upload.onloadstart = (e) =&gt; {</span></a>
<a name="72"><span class="lineNum">      72 </span>                :<span class="lineCov">          5 :                         actual.push('upload loadstart ' + xhr.readyState);</span></a>
<a name="73"><span class="lineNum">      73 </span>                :<span class="lineCov">          5 :                 };</span></a>
<a name="74"><span class="lineNum">      74 </span>                :<span class="lineCov">          5 : </span></a>
<a name="75"><span class="lineNum">      75 </span>           [<span class="branchCov" title="Branch 0 was taken 5 times"> + </span>]:<span class="lineCov">          5 :                 xhr.upload.onprogress = (e) =&gt; {</span></a>
<a name="76"><span class="lineNum">      76 </span>                :<span class="lineCov">          5 :                         // events every 50ms, one final when uploading is done</span></a>
<a name="77"><span class="lineNum">      77 </span>           [<span class="branchNoCov" title="Branch 0 was not taken"> - </span>]:<span class="lineCov">          5 :                         if(xhr.readyState &gt;= xhr.HEADERS_RECEIVED) {</span></a>
<a name="78"><span class="lineNum">      78 </span>                :<span class="lineNoCov">          0 :                                 assert.strictEqual(xhr.status, 200, 'JS never gets to see the 30x status code');</span></a>
<a name="79"><span class="lineNum">      79 </span>                :<span class="lineNoCov">          0 :                         }</span></a>
<a name="80"><span class="lineNum">      80 </span>                :<span class="lineCov">          5 :                         progressFiredReadyState1 = xhr.readyState === xhr.OPENED;</span></a>
<a name="81"><span class="lineNum">      81 </span>                :<span class="lineCov">          5 :                 };</span></a>
<a name="82"><span class="lineNum">      82 </span>                :<span class="lineCov">          5 : </span></a>
<a name="83"><span class="lineNum">      83 </span>           [<span class="branchCov" title="Branch 0 was taken 5 times"> + </span>]:<span class="lineCov">          5 :                 xhr.upload.onloadend = () =&gt; {</span></a>
<a name="84"><span class="lineNum">      84 </span>                :<span class="lineCov">          5 :                         actual.push('upload loadend ' + xhr.readyState);</span></a>
<a name="85"><span class="lineNum">      85 </span>                :<span class="lineCov">          5 :                 };</span></a>
<a name="86"><span class="lineNum">      86 </span>                :<span class="lineCov">          5 : </span></a>
<a name="87"><span class="lineNum">      87 </span>           [<span class="branchCov" title="Branch 0 was taken 5 times"> + </span>]:<span class="lineCov">          5 :                 xhr.onloadstart = () =&gt; {</span></a>
<a name="88"><span class="lineNum">      88 </span>                :<span class="lineCov">          5 :                         actual.push('xhr loadstart ' + xhr.readyState);</span></a>
<a name="89"><span class="lineNum">      89 </span>                :<span class="lineCov">          5 :                 };</span></a>
<a name="90"><span class="lineNum">      90 </span>                :<span class="lineCov">          5 : </span></a>
<a name="91"><span class="lineNum">      91 </span>           [<span class="branchCov" title="Branch 0 was taken 26 times"> + </span>]:<span class="lineCov">          5 :                 xhr.onreadystatechange = () =&gt; {</span></a>
<a name="92"><span class="lineNum">      92 </span>           [<span class="branchCov" title="Branch 0 was taken 21 times"> + </span>]:<span class="lineCov">         26 :                         if(xhr.readyState &gt;= xhr.HEADERS_RECEIVED) {</span></a>
<a name="93"><span class="lineNum">      93 </span>                :<span class="lineCov">         21 :                                 assert.strictEqual(xhr.status, 200, 'JS never gets to see the 30x status code');</span></a>
<a name="94"><span class="lineNum">      94 </span>                :<span class="lineCov">         21 :                         }</span></a>
<a name="95"><span class="lineNum">      95 </span>                :<span class="lineCov">         26 : </span></a>
<a name="96"><span class="lineNum">      96 </span>                :<span class="lineCov">         26 :                         // The UA may fire multiple 'readystatechange' events while in</span></a>
<a name="97"><span class="lineNum">      97 </span>                :<span class="lineCov">         26 :                         // the 'loading' state.</span></a>
<a name="98"><span class="lineNum">      98 </span>                :<span class="lineCov">         26 :                         // https://xhr.spec.whatwg.org/#the-send()-method</span></a>
<a name="99"><span class="lineNum">      99 </span>      [<span class="branchCov" title="Branch 0 was taken 11 times"> + </span>][<span class="branchCov" title="Branch 0 was taken 9 times"> + </span>]:<span class="lineCov">         26 :                         if(xhr.readyState === 3 &amp;&amp; actual[actual.length - 1] === 'xhr onreadystatechange 3') {</span></a>
<a name="100"><span class="lineNum">     100 </span>                :<span class="lineCov">          9 :                                 return;</span></a>
<a name="101"><span class="lineNum">     101 </span>           [<span class="branchCov" title="Branch 0 was taken 17 times"> + </span>]:<span class="lineCov">          9 :                         }</span></a>
<a name="102"><span class="lineNum">     102 </span>                :<span class="lineCov">         17 : </span></a>
<a name="103"><span class="lineNum">     103 </span>                :<span class="lineCov">         17 :                         actual.push('xhr onreadystatechange ' + xhr.readyState);</span></a>
<a name="104"><span class="lineNum">     104 </span>                :<span class="lineCov">          5 :                 };</span></a>
<a name="105"><span class="lineNum">     105 </span>                :<span class="lineCov">          5 : </span></a>
<a name="106"><span class="lineNum">     106 </span>           [<span class="branchCov" title="Branch 0 was taken 5 times"> + </span>]:<span class="lineCov">          5 :                 xhr.onload = (e) =&gt; {</span></a>
<a name="107"><span class="lineNum">     107 </span>                :<span class="lineCov">          5 :                         actual.push('xhr load ' + xhr.readyState);</span></a>
<a name="108"><span class="lineNum">     108 </span>                :<span class="lineCov">          5 :                 };</span></a>
<a name="109"><span class="lineNum">     109 </span>                :<span class="lineCov">          5 : </span></a>
<a name="110"><span class="lineNum">     110 </span>           [<span class="branchCov" title="Branch 0 was taken 5 times"> + </span>]:<span class="lineCov">          5 :                 xhr.onloadend = (e) =&gt; {</span></a>
<a name="111"><span class="lineNum">     111 </span>                :<span class="lineCov">          5 :                         actual.push('xhr loadend ' + xhr.readyState);</span></a>
<a name="112"><span class="lineNum">     112 </span>                :<span class="lineCov">          5 : </span></a>
<a name="113"><span class="lineNum">     113 </span>                :<span class="lineCov">          5 :                         assert(progressFiredReadyState1, 'One progress event should fire on xhr.upload when readyState is 1');</span></a>
<a name="114"><span class="lineNum">     114 </span>                :<span class="lineCov">          5 : </span></a>
<a name="115"><span class="lineNum">     115 </span>                :<span class="lineCov">          5 :                         // Headers will tell us if data was sent when expected</span></a>
<a name="116"><span class="lineNum">     116 </span>           [<span class="branchCov" title="Branch 0 was taken 22 times"> + </span>]:<span class="lineCov">          5 :                         for(let header in expectedHeaders){</span></a>
<a name="117"><span class="lineNum">     117 </span>                :<span class="lineCov">         22 :                                 assert.strictEqual(xhr.getResponseHeader(header), expectedHeaders[header], header);</span></a>
<a name="118"><span class="lineNum">     118 </span>                :<span class="lineCov">         22 :                         }</span></a>
<a name="119"><span class="lineNum">     119 </span>                :<span class="lineCov">          5 : </span></a>
<a name="120"><span class="lineNum">     120 </span>                :<span class="lineCov">          5 :                         assert.deepStrictEqual(actual, expectedEvents, 'events firing in expected order and states');</span></a>
<a name="121"><span class="lineNum">     121 </span>                :<span class="lineCov">          5 : </span></a>
<a name="122"><span class="lineNum">     122 </span>                :<span class="lineCov">          5 :                         if(params.expectedBody)</span></a>
<a name="123"><span class="lineNum">     123 </span>           [<span class="branchCov" title="Branch 0 was taken 2 times"> + </span>]:<span class="lineCov">          5 :                                 assert.strictEqual(xhr.response, params.expectedBody, 'request body was resent');</span></a>
<a name="124"><span class="lineNum">     124 </span>                :<span class="lineCov">          5 : </span></a>
<a name="125"><span class="lineNum">     125 </span>                :<span class="lineCov">          5 :                         xhr.upload.onloadstart = null;</span></a>
<a name="126"><span class="lineNum">     126 </span>                :<span class="lineCov">          5 :                         xhr.upload.onprogress  = null;</span></a>
<a name="127"><span class="lineNum">     127 </span>                :<span class="lineCov">          5 :                         xhr.upload.onloadend   = null;</span></a>
<a name="128"><span class="lineNum">     128 </span>                :<span class="lineCov">          5 : </span></a>
<a name="129"><span class="lineNum">     129 </span>                :<span class="lineCov">          5 :                         xhr.onloadstart = null;</span></a>
<a name="130"><span class="lineNum">     130 </span>                :<span class="lineCov">          5 :                         xhr.onprogress  = null;</span></a>
<a name="131"><span class="lineNum">     131 </span>                :<span class="lineCov">          5 :                         xhr.onload      = null;</span></a>
<a name="132"><span class="lineNum">     132 </span>                :<span class="lineCov">          5 :                         xhr.onloadend   = null;</span></a>
<a name="133"><span class="lineNum">     133 </span>                :<span class="lineCov">          5 :                 };</span></a>
<a name="134"><span class="lineNum">     134 </span>                :<span class="lineCov">          5 : </span></a>
<a name="135"><span class="lineNum">     135 </span>                :<span class="lineCov">          5 :                 xhr.open('POST', `${activeURL}/redirect.py?location=/xhr/resources/content.py&amp;code=${params.code}`);</span></a>
<a name="136"><span class="lineNum">     136 </span>                :<span class="lineCov">          5 :                 xhr.send(params.body);</span></a>
<a name="137"><span class="lineNum">     137 </span>                :<span class="lineCov">          5 :         }</span></a>
<a name="138"><span class="lineNum">     138 </span>                :<span class="lineCov">          1 : }</span></a>
<a name="139"><span class="lineNum">     139 </span>                :<span class="lineCov">          1 : ﻿</span></a>
<a name="140"><span class="lineNum">     140 </span>                :<span class="lineCov">          1 : /*</span></a>
<a name="141"><span class="lineNum">     141 </span>                :<span class="lineCov">          1 :  * send-redirect-post-upload.htm</span></a>
<a name="142"><span class="lineNum">     142 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="143"><span class="lineNum">     143 </span>                :<span class="lineCov">          1 : </span></a>
<a name="144"><span class="lineNum">     144 </span>                :<span class="lineCov">          1 : &lt;!DOCTYPE html&gt;</span></a>
<a name="145"><span class="lineNum">     145 </span>                :<span class="lineCov">          1 : &lt;html&gt;</span></a>
<a name="146"><span class="lineNum">     146 </span>                :<span class="lineCov">          1 : &lt;head&gt;</span></a>
<a name="147"><span class="lineNum">     147 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#handler-xhr-onprogress&quot; data-tested-assertations=&quot;../..&quot; /&gt;</span></a>
<a name="148"><span class="lineNum">     148 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#event-xhr-progress&quot; data-tested-assertations=&quot;../..&quot; /&gt;</span></a>
<a name="149"><span class="lineNum">     149 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#the-send()-method&quot; data-tested-assertations=&quot;following::dt[@id=&quot;dom-xmlhttprequest-send-bodyinit&quot;]/following::dd[1]/p[2] following::ol[1]/li[9]//li[1] following::ol[1]/li[9]//li[2]&quot; /&gt;</span></a>
<a name="150"><span class="lineNum">     150 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://fetch.spec.whatwg.org/#http-fetch&quot; data-tested-assertations=&quot;following::ol[1]/li[6]/dl/dd[1]//dd[3]&quot; /&gt;</span></a>
<a name="151"><span class="lineNum">     151 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://fetch.spec.whatwg.org/#concept-http-redirect-fetch&quot; data-tested-assertations=&quot;following::li[16]&quot; /&gt;</span></a>
<a name="152"><span class="lineNum">     152 </span>                :<span class="lineCov">          1 :     &lt;script src=&quot;/resources/testharness.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="153"><span class="lineNum">     153 </span>                :<span class="lineCov">          1 :     &lt;script src=&quot;/resources/testharnessreport.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="154"><span class="lineNum">     154 </span>                :<span class="lineCov">          1 :     &lt;title&gt;XMLHttpRequest: The send() method: POSTing to URL that redirects&lt;/title&gt;</span></a>
<a name="155"><span class="lineNum">     155 </span>                :<span class="lineCov">          1 : &lt;/head&gt;</span></a>
<a name="156"><span class="lineNum">     156 </span>                :<span class="lineCov">          1 : </span></a>
<a name="157"><span class="lineNum">     157 </span>                :<span class="lineCov">          1 : &lt;body&gt;</span></a>
<a name="158"><span class="lineNum">     158 </span>                :<span class="lineCov">          1 :     &lt;div id=&quot;log&quot;&gt;&lt;/div&gt;</span></a>
<a name="159"><span class="lineNum">     159 </span>                :<span class="lineCov">          1 : </span></a>
<a name="160"><span class="lineNum">     160 </span>                :<span class="lineCov">          1 :     &lt;script type=&quot;text/javascript&quot;&gt;</span></a>
<a name="161"><span class="lineNum">     161 </span>                :<span class="lineCov">          1 :     function testRedirectPost(params) {</span></a>
<a name="162"><span class="lineNum">     162 </span>                :<span class="lineCov">          1 :         var test = async_test(document.title + &quot; (&quot; + params.name + &quot;)&quot;);</span></a>
<a name="163"><span class="lineNum">     163 </span>                :<span class="lineCov">          1 :         var actual = [];</span></a>
<a name="164"><span class="lineNum">     164 </span>                :<span class="lineCov">          1 :         // We check upload.onprogress with a boolean because it *might* fire more than once</span></a>
<a name="165"><span class="lineNum">     165 </span>                :<span class="lineCov">          1 :         var progressFiredReadyState1 = false;</span></a>
<a name="166"><span class="lineNum">     166 </span>                :<span class="lineCov">          1 : </span></a>
<a name="167"><span class="lineNum">     167 </span>                :<span class="lineCov">          1 :         var expectedHeaders, expectedEvents;</span></a>
<a name="168"><span class="lineNum">     168 </span>                :<span class="lineCov">          1 : </span></a>
<a name="169"><span class="lineNum">     169 </span>                :<span class="lineCov">          1 :         // 307 redirects should resend the POST data, and events and headers will be a little different..</span></a>
<a name="170"><span class="lineNum">     170 </span>                :<span class="lineCov">          1 :         if(params.expectResendPost) {</span></a>
<a name="171"><span class="lineNum">     171 </span>                :<span class="lineCov">          1 :             expectedHeaders = {</span></a>
<a name="172"><span class="lineNum">     172 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Content-Length&quot;: &quot;12000&quot;,</span></a>
<a name="173"><span class="lineNum">     173 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Content-Type&quot;: &quot;text/plain;charset=UTF-8&quot;,</span></a>
<a name="174"><span class="lineNum">     174 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Method&quot;: &quot;POST&quot;,</span></a>
<a name="175"><span class="lineNum">     175 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Query&quot;: &quot;NO&quot;,</span></a>
<a name="176"><span class="lineNum">     176 </span>                :<span class="lineCov">          1 :                 &quot;Content-Length&quot;: &quot;12000&quot;</span></a>
<a name="177"><span class="lineNum">     177 </span>                :<span class="lineCov">          1 :             }</span></a>
<a name="178"><span class="lineNum">     178 </span>                :<span class="lineCov">          1 :             expectedEvents = [</span></a>
<a name="179"><span class="lineNum">     179 </span>                :<span class="lineCov">          1 :                 &quot;xhr onreadystatechange 1&quot;,</span></a>
<a name="180"><span class="lineNum">     180 </span>                :<span class="lineCov">          1 :                 &quot;xhr loadstart 1&quot;,</span></a>
<a name="181"><span class="lineNum">     181 </span>                :<span class="lineCov">          1 :                 &quot;upload loadstart 1&quot;,</span></a>
<a name="182"><span class="lineNum">     182 </span>                :<span class="lineCov">          1 :                 &quot;upload loadend 1&quot;,</span></a>
<a name="183"><span class="lineNum">     183 </span>                :<span class="lineCov">          1 :                 &quot;xhr onreadystatechange 2&quot;,</span></a>
<a name="184"><span class="lineNum">     184 </span>                :<span class="lineCov">          1 :                 &quot;xhr onreadystatechange 3&quot;,</span></a>
<a name="185"><span class="lineNum">     185 </span>                :<span class="lineCov">          1 :                 &quot;xhr onreadystatechange 4&quot;,</span></a>
<a name="186"><span class="lineNum">     186 </span>                :<span class="lineCov">          1 :                 &quot;xhr load 4&quot;,</span></a>
<a name="187"><span class="lineNum">     187 </span>                :<span class="lineCov">          1 :                 &quot;xhr loadend 4&quot;</span></a>
<a name="188"><span class="lineNum">     188 </span>                :<span class="lineCov">          1 :             ];</span></a>
<a name="189"><span class="lineNum">     189 </span>                :<span class="lineCov">          1 :         } else {</span></a>
<a name="190"><span class="lineNum">     190 </span>                :<span class="lineCov">          1 :             // setting the right expectations for POST resent as GET without request body</span></a>
<a name="191"><span class="lineNum">     191 </span>                :<span class="lineCov">          1 :             expectedHeaders = {</span></a>
<a name="192"><span class="lineNum">     192 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Content-Length&quot;: &quot;NO&quot;,</span></a>
<a name="193"><span class="lineNum">     193 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Content-Type&quot;: &quot;NO&quot;,</span></a>
<a name="194"><span class="lineNum">     194 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Method&quot;: &quot;GET&quot;,</span></a>
<a name="195"><span class="lineNum">     195 </span>                :<span class="lineCov">          1 :                 &quot;X-Request-Query&quot;: &quot;NO&quot;</span></a>
<a name="196"><span class="lineNum">     196 </span>                :<span class="lineCov">          1 :             }</span></a>
<a name="197"><span class="lineNum">     197 </span>                :<span class="lineCov">          1 :             expectedEvents = [</span></a>
<a name="198"><span class="lineNum">     198 </span>                :<span class="lineCov">          1 :                 &quot;xhr onreadystatechange 1&quot;,</span></a>
<a name="199"><span class="lineNum">     199 </span>                :<span class="lineCov">          1 :                 &quot;xhr loadstart 1&quot;,</span></a>
<a name="200"><span class="lineNum">     200 </span>                :<span class="lineCov">          1 :                 &quot;upload loadstart 1&quot;,</span></a>
<a name="201"><span class="lineNum">     201 </span>                :<span class="lineCov">          1 :                 &quot;upload loadend 1&quot;,</span></a>
<a name="202"><span class="lineNum">     202 </span>                :<span class="lineCov">          1 :                 &quot;xhr onreadystatechange 2&quot;,</span></a>
<a name="203"><span class="lineNum">     203 </span>                :<span class="lineCov">          1 :                 // we expect no onreadystatechange readyState=3 event because there is no loading content</span></a>
<a name="204"><span class="lineNum">     204 </span>                :<span class="lineCov">          1 :                 &quot;xhr onreadystatechange 4&quot;,</span></a>
<a name="205"><span class="lineNum">     205 </span>                :<span class="lineCov">          1 :                 &quot;xhr load 4&quot;,</span></a>
<a name="206"><span class="lineNum">     206 </span>                :<span class="lineCov">          1 :                 &quot;xhr loadend 4&quot;</span></a>
<a name="207"><span class="lineNum">     207 </span>                :<span class="lineCov">          1 :             ];</span></a>
<a name="208"><span class="lineNum">     208 </span>                :<span class="lineCov">          1 :         }</span></a>
<a name="209"><span class="lineNum">     209 </span>                :<span class="lineCov">          1 :         // Override expectations if provided.</span></a>
<a name="210"><span class="lineNum">     210 </span>                :<span class="lineCov">          1 :         if(params.expectedContentType)</span></a>
<a name="211"><span class="lineNum">     211 </span>                :<span class="lineCov">          1 :           expectedHeaders[&quot;X-Request-Content-Type&quot;] = params.expectedContentType;</span></a>
<a name="212"><span class="lineNum">     212 </span>                :<span class="lineCov">          1 : </span></a>
<a name="213"><span class="lineNum">     213 </span>                :<span class="lineCov">          1 :         test.step(function()</span></a>
<a name="214"><span class="lineNum">     214 </span>                :<span class="lineCov">          1 :         {</span></a>
<a name="215"><span class="lineNum">     215 </span>                :<span class="lineCov">          1 :             var xhr = new XMLHttpRequest();</span></a>
<a name="216"><span class="lineNum">     216 </span>                :<span class="lineCov">          1 : </span></a>
<a name="217"><span class="lineNum">     217 </span>                :<span class="lineCov">          1 :             xhr.upload.onloadstart = test.step_func(function(e) {</span></a>
<a name="218"><span class="lineNum">     218 </span>                :<span class="lineCov">          1 :                 actual.push(&quot;upload loadstart &quot; + xhr.readyState);</span></a>
<a name="219"><span class="lineNum">     219 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="220"><span class="lineNum">     220 </span>                :<span class="lineCov">          1 :             xhr.upload.onprogress = test.step_func(function(e) {</span></a>
<a name="221"><span class="lineNum">     221 </span>                :<span class="lineCov">          1 :                 // events every 50ms, one final when uploading is done</span></a>
<a name="222"><span class="lineNum">     222 </span>                :<span class="lineCov">          1 :                 if(xhr.readyState &gt;= xhr.HEADERS_RECEIVED) {</span></a>
<a name="223"><span class="lineNum">     223 </span>                :<span class="lineCov">          1 :                     assert_equals(xhr.status, 200, &quot;JS never gets to see the 30x status code&quot;);</span></a>
<a name="224"><span class="lineNum">     224 </span>                :<span class="lineCov">          1 :                 }</span></a>
<a name="225"><span class="lineNum">     225 </span>                :<span class="lineCov">          1 :                 progressFiredReadyState1 = xhr.readyState === xhr.OPENED;</span></a>
<a name="226"><span class="lineNum">     226 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="227"><span class="lineNum">     227 </span>                :<span class="lineCov">          1 :             xhr.upload.onloadend = test.step_func(function() {</span></a>
<a name="228"><span class="lineNum">     228 </span>                :<span class="lineCov">          1 :                 actual.push(&quot;upload loadend &quot; + xhr.readyState);</span></a>
<a name="229"><span class="lineNum">     229 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="230"><span class="lineNum">     230 </span>                :<span class="lineCov">          1 :             xhr.onloadstart = test.step_func(function() {</span></a>
<a name="231"><span class="lineNum">     231 </span>                :<span class="lineCov">          1 :                 actual.push(&quot;xhr loadstart &quot; + xhr.readyState);</span></a>
<a name="232"><span class="lineNum">     232 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="233"><span class="lineNum">     233 </span>                :<span class="lineCov">          1 :             xhr.onreadystatechange = test.step_func(function() {</span></a>
<a name="234"><span class="lineNum">     234 </span>                :<span class="lineCov">          1 :                 if(xhr.readyState &gt;= xhr.HEADERS_RECEIVED) {</span></a>
<a name="235"><span class="lineNum">     235 </span>                :<span class="lineCov">          1 :                     assert_equals(xhr.status, 200, &quot;JS never gets to see the 30x status code&quot;);</span></a>
<a name="236"><span class="lineNum">     236 </span>                :<span class="lineCov">          1 :                 }</span></a>
<a name="237"><span class="lineNum">     237 </span>                :<span class="lineCov">          1 : </span></a>
<a name="238"><span class="lineNum">     238 </span>                :<span class="lineCov">          1 :                 // The UA may fire multiple &quot;readystatechange&quot; events while in</span></a>
<a name="239"><span class="lineNum">     239 </span>                :<span class="lineCov">          1 :                 // the &quot;loading&quot; state.</span></a>
<a name="240"><span class="lineNum">     240 </span>                :<span class="lineCov">          1 :                 // https://xhr.spec.whatwg.org/#the-send()-method</span></a>
<a name="241"><span class="lineNum">     241 </span>                :<span class="lineCov">          1 :                 if (xhr.readyState === 3 &amp;&amp; actual[actual.length - 1] === &quot;xhr onreadystatechange 3&quot;) {</span></a>
<a name="242"><span class="lineNum">     242 </span>                :<span class="lineCov">          1 :                     return;</span></a>
<a name="243"><span class="lineNum">     243 </span>                :<span class="lineCov">          1 :                 }</span></a>
<a name="244"><span class="lineNum">     244 </span>                :<span class="lineCov">          1 : </span></a>
<a name="245"><span class="lineNum">     245 </span>                :<span class="lineCov">          1 :                 actual.push(&quot;xhr onreadystatechange &quot; + xhr.readyState);</span></a>
<a name="246"><span class="lineNum">     246 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="247"><span class="lineNum">     247 </span>                :<span class="lineCov">          1 :             xhr.onload = test.step_func(function(e)</span></a>
<a name="248"><span class="lineNum">     248 </span>                :<span class="lineCov">          1 :             {</span></a>
<a name="249"><span class="lineNum">     249 </span>                :<span class="lineCov">          1 :                 actual.push(&quot;xhr load &quot; + xhr.readyState);</span></a>
<a name="250"><span class="lineNum">     250 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="251"><span class="lineNum">     251 </span>                :<span class="lineCov">          1 :             xhr.onloadend = test.step_func(function(e)</span></a>
<a name="252"><span class="lineNum">     252 </span>                :<span class="lineCov">          1 :             {</span></a>
<a name="253"><span class="lineNum">     253 </span>                :<span class="lineCov">          1 :                 actual.push(&quot;xhr loadend &quot; + xhr.readyState);</span></a>
<a name="254"><span class="lineNum">     254 </span>                :<span class="lineCov">          1 : </span></a>
<a name="255"><span class="lineNum">     255 </span>                :<span class="lineCov">          1 :                 assert_true(progressFiredReadyState1, &quot;One progress event should fire on xhr.upload when readyState is 1&quot;);</span></a>
<a name="256"><span class="lineNum">     256 </span>                :<span class="lineCov">          1 : </span></a>
<a name="257"><span class="lineNum">     257 </span>                :<span class="lineCov">          1 :                 // Headers will tell us if data was sent when expected</span></a>
<a name="258"><span class="lineNum">     258 </span>                :<span class="lineCov">          1 :                 for(var header in expectedHeaders) {</span></a>
<a name="259"><span class="lineNum">     259 </span>                :<span class="lineCov">          1 :                     assert_equals(xhr.getResponseHeader(header), expectedHeaders[header], header);</span></a>
<a name="260"><span class="lineNum">     260 </span>                :<span class="lineCov">          1 :                 }</span></a>
<a name="261"><span class="lineNum">     261 </span>                :<span class="lineCov">          1 : </span></a>
<a name="262"><span class="lineNum">     262 </span>                :<span class="lineCov">          1 :                 assert_array_equals(actual, expectedEvents, &quot;events firing in expected order and states&quot;);</span></a>
<a name="263"><span class="lineNum">     263 </span>                :<span class="lineCov">          1 :                 if (params.expectedBody)</span></a>
<a name="264"><span class="lineNum">     264 </span>                :<span class="lineCov">          1 :                   assert_equals(xhr.response, params.expectedBody, 'request body was resent');</span></a>
<a name="265"><span class="lineNum">     265 </span>                :<span class="lineCov">          1 :                 test.done();</span></a>
<a name="266"><span class="lineNum">     266 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="267"><span class="lineNum">     267 </span>                :<span class="lineCov">          1 : </span></a>
<a name="268"><span class="lineNum">     268 </span>                :<span class="lineCov">          1 :             xhr.open(&quot;POST&quot;, &quot;./resources/redirect.py?location=content.py&amp;code=&quot; + params.code, true);</span></a>
<a name="269"><span class="lineNum">     269 </span>                :<span class="lineCov">          1 :             xhr.send(params.body);</span></a>
<a name="270"><span class="lineNum">     270 </span>                :<span class="lineCov">          1 :         });</span></a>
<a name="271"><span class="lineNum">     271 </span>                :<span class="lineCov">          1 :     }</span></a>
<a name="272"><span class="lineNum">     272 </span>                :<span class="lineCov">          1 : </span></a>
<a name="273"><span class="lineNum">     273 </span>                :<span class="lineCov">          1 :     const stringBody = &quot;Test Message&quot;.repeat(1000);</span></a>
<a name="274"><span class="lineNum">     274 </span>                :<span class="lineCov">          1 :     const blobBody = new Blob(new Array(1000).fill(&quot;Test Message&quot;));</span></a>
<a name="275"><span class="lineNum">     275 </span>                :<span class="lineCov">          1 : </span></a>
<a name="276"><span class="lineNum">     276 </span>                :<span class="lineCov">          1 :     testRedirectPost({name: &quot;301&quot;, code: 301, expectResendPost: false, body: stringBody});</span></a>
<a name="277"><span class="lineNum">     277 </span>                :<span class="lineCov">          1 :     testRedirectPost({name: &quot;302&quot;, code: 302, expectResendPost: false, body: stringBody});</span></a>
<a name="278"><span class="lineNum">     278 </span>                :<span class="lineCov">          1 :     testRedirectPost({name: &quot;303&quot;, code: 303, expectResendPost: false, body: stringBody});</span></a>
<a name="279"><span class="lineNum">     279 </span>                :<span class="lineCov">          1 :     testRedirectPost({name: &quot;307 (string)&quot;, code: 307, expectResendPost: true,  body: stringBody, expectedBody: stringBody });</span></a>
<a name="280"><span class="lineNum">     280 </span>                :<span class="lineCov">          1 :     testRedirectPost({name: &quot;307 (blob)&quot;, code: 307, expectResendPost: true, body: blobBody, expectedBody: stringBody, expectedContentType: &quot;NO&quot; });</span></a>
<a name="281"><span class="lineNum">     281 </span>                :<span class="lineCov">          1 :     &lt;/script&gt;</span></a>
<a name="282"><span class="lineNum">     282 </span>                :<span class="lineCov">          1 : &lt;/body&gt;</span></a>
<a name="283"><span class="lineNum">     283 </span>                :<span class="lineCov">          1 : &lt;/html&gt;</span></a>
<a name="284"><span class="lineNum">     284 </span>                :<span class="lineCov">          1 : </span></a>
<a name="285"><span class="lineNum">     285 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="286"><span class="lineNum">     286 </span>                :<span class="lineCov">          1 :  * send-redirect-post-upload.htm</span></a>
<a name="287"><span class="lineNum">     287 </span>                :<span class="lineCov">          1 :  */</span></a>
</pre>
      </td>
    </tr>
  </table>
  <br>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
  </table>
  <br>

</body>
</html>
